% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GMM.R
\name{plot_GMM}
\alias{plot_GMM}
\title{Plots Components from Gaussian Mixture Models}
\usage{
plot_GMM(m, k=NULL)
}
\arguments{
\item{m}{a \code{mixEM} class object corresponding with the fit GMM}

\item{k}{the number of components specified in the GMM, \code{m}}
}
\description{
Generates a plot of densities with overlaid components form a Gaussian mixture model (GMM)
}
\details{
Uses ggplot2 graphics to plot data densities with overlaid components from \code{mixEM} objects, which are GMM's fit using the \code{mixtools} package.

Note: Users must enter the same component value, \code{k}, in the \code{plot_GMM} function, as that which was specified in the original GMM specification.
}
\examples{
mixmdl <- mixtools::normalmixEM(faithful$waiting, k = 2)

plot_GMM(mixmdl, 2)
}
